add transpose matrix
